gestao_clientes/requirements-dev.txt
certifi==2020.6.20
chardet==3.0.4
defusedxml==0.6.0
dj-database-url==0.4.2
dj-static==0.0.6
Django==2.0.1
django-admin-interface==0.12.2
django-allauth==0.42.0
django-bootstrap-form==3.3
django-colorfield==0.3.1
django-debug-toolbar==2.2
django-flat-responsive==2.0
django-flat-theme==1.1.4
idna==2.10
oauthlib==3.1.0
Pillow==7.1.2
python-decouple==3.1
python3-openid==3.2.0
pytz==2017.3
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1
static3==0.7.0
urllib3==1.25.9
